Remove `RequestUtils`
